feat(frontend): move api credential set action to own endpoint
- removes the action from the index. the intention is to
expose the remix server port over docker and  call this
from the MASE to set the api credentials on start
